Adding --clean flag to reduce build times on CircleCI

+++ test/setupDev.sh
@@ -5,18 +5,20 @@ green () {
     echo -e "\033[1;32m== $1 \033[0m"
 }
 
-green 'Removing old node_modules dirs if present'
-if [ -d ../client/node_modules ]; then
-  echo Removing client/node_modules
-  rm -r ../client/node_modules
-fi
-if [ -d ../server/node_modules ]; then
+if [ $1 == '--clean' ]; then
+ green 'Removing old node_modules dirs if present'
+ if [ -d ../client/node_modules ]; then
+   echo Removing client/node_modules
+   rm -r ../client/node_modules
+ fi
+ if [ -d ../server/node_modules ]; then
   echo Removing server/node_modules
   rm -r ../server/node_modules
-fi
-if [ -d ../cli/node_modules ]; then
+ fi
+ if [ -d ../cli/node_modules ]; then
   echo Removing cli/node_modules
   rm -r ../cli/node_modules
+ fi
 fi
 
 pushd ../client
